move data type edit view into data types folder
